阿发

+++ WlanOrderAutomatic/src/com/jobsAutomatic/service/util/UUIDGenerator.java
@@ -70,7 +70,7 @@ public abstract class UUIDGenerator{
 	public static int toInt( byte[] bytes ) {
 		int result = 0;
 		for (int i=0; i<4; i++) {
-			result = ( result << 8 ) - Byte.MIN_VALUE + (int) bytes[i];
+			result = ( result << 8 ) - Byte.MIN_VALUE + bytes[i];
 		}
 		return result;
 	}
